requirements.txt
datasets==4.4.1
evaluate==0.4.6
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0
torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
